Add e2e test for PPR navigations
Creates an e2e test case for prefetching and navigating when PPR is
enabled. While I was writing this test, I noticed that it's currently
pretty difficult to test all the scenarios that PPR is designed to
handle, so I gave special attention to setting up a testing strategy
that I hope will make this easier going forward. The overall pattern is
based on how we've been testing concurrent rendering features in the
React repo for many years:

- In the e2e test, spin up an HTTP server for responding to requests
sent by the test app. This simulates the data service that would be used
in a real Next.js application, whether it's direct db access, an ORM, or
a higher-level data access layer. The e2e test can observe when
individual requests are received, and control the timing of when the
data is fulfilled, without needing to mock any lower level I/O. (We're
already using a similar pattern to test fetch deduping:
https://github.com/vercel/next.js/blob/a3616d33edc40e7717b258f2636f280e30a3bccb/test/e2e/app-dir/app-fetch-deduping/app-fetch-deduping.test.ts#L8-L29)
- Each time a request is received, write to an event log. Then assert on
the result of the log at different points throughout the test. This
helps catch subtle mistakes where the order of events is not expected,
or the same event happens more than it should.

(I wrote some test helpers, but to avoid early abstraction, I've
intentionally not moved them into a separate module.)