Move `head` to root Cache Node
The Cache Node type has a `head` property. The head conceptually belongs
to a single segment — the currently active page — but there's only one
per router state, and it's rendered by the root component (AppRouter).

Currently, AppRouter will do a depth-first search of the Cache Node
tree to find the page segment and access its `head` property. But since
the `head` is delivered by the same server response as the rest of the
Cache Node data, we can put the head data on the root Cache Node using
the same function — fillLazyItemsTillLeafWithHead — that fills the
Cache Nodes.

Really it should go on the reducer state object instead, so that we
don't have to add an extra field to the Cache Nodes. We can do in a
follow up where we also rename this fillLazyItemsTillLeafWithHead, since
it's more about filling the cache with data as opposed to setting the
`head` property. OTOH, this will likely be replaced in the PPR
implementation anyway, which is why I'm leaving it as-is for now.