[PPR] Skip fetch when navigating to fully static page
When navigating to a fully static page with no dynamic holes, if the
RSC data was already prefetched, we don't need to do an additional
fetch. We can treat the prefetched data as complete.

Ideally we should be able to do this per-segment, rather than per-page,
but the static payload currently does not include information about
which segments are fully static. In the future, we will add the ability
to fetch (and prefetch) at the segment level.