[Segment Cache] Add chars to cache key allowlist
Adds (, ), ~, and . to the allowlist of characters that don't need to
be base64url encoded when creating a segment cache key, since these are
all URL- and filesystem-safe.

Doesn't affect any external behavior. It just makes certain segment
paths slightly more compact, and easier to debug when inspecting the
build output.