Add parsedUrl to AppRenderContext
Adds parsedUrl to the AppRenderContext. We were already passing
the `query`, but now I need to the corresponding pathname, too.

This is the "post-rewrite" URL. It's the one that is observed by
components via the `params` and `searchParams` props. If there is a
rewrite, it corresponds to the `x-nextjs-rewritten-path` and
`x-nextjs-rewritten-query` headers.

I've also added this to the initial RSC payload so it can be sent to
the client during the initial render. It doesn't need to be sent during
a client navigation/refresh/revalidation, because the client will read
the information from the response headers instead.