Pass RouteTree into navigation function
RouteTree is the client version of FlightRouterState. It's the same
representation of the route tree, but it structured for optimized
lookups of the client cache. The plan is the make this the primary/only
type used for dealing with routes on the client; FlightRouterState will
be used a transport format only.

Using the same type accessing the cache during both prefetches and
navigations makes it less likely for the behavior between the two
paths to drift, too. Now that we've deleted the old, pre-Segment
Cache navigation implementation, we should generally try to unify the
data structures wherever possible.